Added basic status effect system. Includes scriptable objects, and a few custom-made icons. Only poison actively works right now (would work for all damage-dealing effects if I added more), have to work on a UI design, and adding buff-related effects.

+++ Assets/Scripts/Battle System/TurnManager.cs
@@ -33,7 +33,7 @@ public class TurnManager
 
         turnOrder.Sort((a, b) =>
         {
-            return a.GetComponent<EntityStats>().delay - b.GetComponent<EntityStats>().delay;
+            return (int)a.GetComponent<EntityStats>().delay - (int)b.GetComponent<EntityStats>().delay;
         });
     }
